+++ logger/proto.go
@@ -15,15 +15,21 @@
 package logger
 
 import (
+	"bytes"
 	"encoding/base64"
 	"fmt"
 	"strconv"
+	"text/template"
 
 	"go.uber.org/zap/zapcore"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protoreflect"
+	"google.golang.org/protobuf/types/dynamicpb"
+
+	"github.com/puzpuzpuz/xsync/v3"
 
 	"github.com/livekit/protocol/livekit/logger"
+	"github.com/livekit/protocol/utils/must"
 )
 
 func Proto(val proto.Message) zapcore.ObjectMarshaler {
@@ -42,14 +48,22 @@ type protoMarshaller struct {
 }
 
 func (p protoMarshaller) MarshalLogObject(e zapcore.ObjectEncoder) error {
+	if !p.m.IsValid() {
+		return nil
+	}
+
 	fields := p.m.Descriptor().Fields()
 	for i := 0; i < fields.Len(); i++ {
 		f := fields.Get(i)
 		k := f.JSONName()
 		v := p.m.Get(f)
 
+		if protoFieldIsZero(f, v) {
+			continue
+		}
+
 		if proto.HasExtension(f.Options(), logger.E_Redact) {
-			e.AddString(k, "<redacted>")
+			e.AddString(k, marshalRedacted(f, v))
 			continue
 		}
 
@@ -125,35 +139,21 @@ func (p protoListMarshaller) MarshalLogArray(e zapcore.ArrayEncoder) error {
 func marshalProtoField(k string, f protoreflect.FieldDescriptor, v protoreflect.Value, e zapcore.ObjectEncoder) {
 	switch f.Kind() {
 	case protoreflect.BoolKind:
-		if b := v.Bool(); b {
-			e.AddBool(k, b)
-		}
+		e.AddBool(k, v.Bool())
 	case protoreflect.EnumKind:
 		e.AddString(k, marshalProtoEnum(f, v))
 	case protoreflect.Int32Kind, protoreflect.Int64Kind, protoreflect.Sint32Kind, protoreflect.Sint64Kind, protoreflect.Sfixed32Kind, protoreflect.Sfixed64Kind:
-		if n := v.Int(); n != 0 {
-			e.AddInt64(k, n)
-		}
+		e.AddInt64(k, v.Int())
 	case protoreflect.Uint32Kind, protoreflect.Uint64Kind, protoreflect.Fixed32Kind, protoreflect.Fixed64Kind:
-		if n := v.Uint(); n != 0 {
-			e.AddUint64(k, n)
-		}
+		e.AddUint64(k, v.Uint())
 	case protoreflect.FloatKind, protoreflect.DoubleKind:
-		if n := v.Float(); n != 0 {
-			e.AddFloat64(k, n)
-		}
+		e.AddFloat64(k, v.Float())
 	case protoreflect.StringKind:
-		if s := v.String(); s != "" {
-			e.AddString(k, s)
-		}
+		e.AddString(k, v.String())
 	case protoreflect.BytesKind:
-		if b := v.Bytes(); len(b) != 0 {
-			e.AddString(k, marshalProtoBytes(b))
-		}
+		e.AddString(k, marshalProtoBytes(v.Bytes()))
 	case protoreflect.MessageKind:
-		if m := v.Message(); m.IsValid() {
-			e.AddObject(k, protoMarshaller{m})
-		}
+		e.AddObject(k, protoMarshaller{v.Message()})
 	}
 }
 
@@ -182,4 +182,84 @@ func marshalProtoBytes(b []byte) string {
 	default:
 		return fmt.Sprintf("%s... (%.2fGB)", s, float64(n)/float64(1<<30))
 	}
+}
+
+var redactTemplates = xsync.NewMapOf[string, *template.Template]()
+
+func marshalRedacted(f protoreflect.FieldDescriptor, v protoreflect.Value) string {
+	if !proto.HasExtension(f.Options(), logger.E_RedactFormat) {
+		return "<redacted>"
+	}
+
+	text := proto.GetExtension(f.Options(), logger.E_RedactFormat).(string)
+	tpl, _ := redactTemplates.LoadOrCompute(text, func() *template.Template {
+		return template.Must(template.New("format").Parse(text))
+	})
+
+	var b bytes.Buffer
+	must.Do(tpl.Execute(&b, redactTemplateData{f, v}))
+	return b.String()
+}
+
+type redactTemplateData struct {
+	f protoreflect.FieldDescriptor
+	v protoreflect.Value
+}
+
+func (d redactTemplateData) TextName() string {
+	return d.f.TextName()
+}
+
+func (d redactTemplateData) Size() string {
+	msg := dynamicpb.NewMessage(d.f.ContainingMessage())
+
+	switch {
+	case d.f.IsList():
+		dst := msg.Mutable(d.f).List()
+		src := d.v.List()
+		for i := 0; i < src.Len(); i++ {
+			dst.Append(src.Get(i))
+		}
+	case d.f.IsMap():
+		dst := msg.Mutable(d.f).Map()
+		src := d.v.Map()
+		src.Range(func(k protoreflect.MapKey, v protoreflect.Value) bool {
+			dst.Set(k, v)
+			return true
+		})
+	default:
+		msg.Set(d.f, d.v)
+	}
+
+	return strconv.Itoa(proto.Size(msg.Interface()))
+}
+
+func protoFieldIsZero(f protoreflect.FieldDescriptor, v protoreflect.Value) bool {
+	if f.IsList() {
+		l := v.List()
+		return l == nil || l.Len() == 0
+	}
+	if f.IsMap() {
+		m := v.Map()
+		return m == nil || m.Len() == 0
+	}
+	switch f.Kind() {
+	case protoreflect.BoolKind:
+		return !v.Bool()
+	case protoreflect.EnumKind:
+		return false
+	case protoreflect.Int32Kind, protoreflect.Int64Kind, protoreflect.Sint32Kind, protoreflect.Sint64Kind, protoreflect.Sfixed32Kind, protoreflect.Sfixed64Kind:
+		return v.Int() == 0
+	case protoreflect.Uint32Kind, protoreflect.Uint64Kind, protoreflect.Fixed32Kind, protoreflect.Fixed64Kind:
+		return v.Uint() == 0
+	case protoreflect.FloatKind, protoreflect.DoubleKind:
+		return v.Float() == 0
+	case protoreflect.StringKind:
+		return v.String() == ""
+	case protoreflect.BytesKind:
+		return len(v.Bytes()) == 0
+	case protoreflect.MessageKind:
+		return !v.Message().IsValid()
+	}
+	return true
 }