+++ observability/reporter.go
@@ -2,6 +2,7 @@ package observability
 
 import (
 	"github.com/livekit/protocol/observability/agentsobs"
+	"github.com/livekit/protocol/observability/egressobs"
 	"github.com/livekit/protocol/observability/gatewayobs"
 	"github.com/livekit/protocol/observability/roomobs"
 	"github.com/livekit/protocol/observability/telephonyobs"
@@ -14,6 +15,9 @@ type Reporter interface {
 	Agent() agentsobs.Reporter
 	Gateway() gatewayobs.Reporter
 	Telephony() telephonyobs.Reporter
+	Connector() any // any is a placeholder for the connector type
+	Egress() egressobs.Reporter
+	Ingress() any
 	Close()
 }
 
@@ -39,4 +43,17 @@ func (reporter) Telephony() telephonyobs.Reporter {
 	return telephonyobs.NewNoopReporter()
 }
 
-func (reporter) Close() {}
+func (reporter) Connector() any {
+	return nil
+}
+
+func (reporter) Egress() egressobs.Reporter {
+	return egressobs.NewNoopReporter()
+}
+
+func (reporter) Ingress() any {
+	return nil
+}
+
+func (reporter) Close() {
+}