+++ packages/javascript/CHANGELOG.md
@@ -1,5 +1,81 @@
 # @livekit/protocol
 
+## 1.44.0
+
+### Patch Changes
+
+- Add auto_subscribe_data_track to StartSession. - [#1366](https://github.com/livekit/protocol/pull/1366) ([@boks1971](https://github.com/boks1971))
+
+- add helper to propagate psrpc request timeout - [#1361](https://github.com/livekit/protocol/pull/1361) ([@paulwe](https://github.com/paulwe))
+
+- skip logging redacted fields with zero values - [#1338](https://github.com/livekit/protocol/pull/1338) ([@paulwe](https://github.com/paulwe))
+
+- redact metadata in agent protos - [#1344](https://github.com/livekit/protocol/pull/1344) ([@paulwe](https://github.com/paulwe))
+
+- Add featureinfo nested field for reporting noise cancellation feature specific metadata - [#1367](https://github.com/livekit/protocol/pull/1367) ([@1egoman](https://github.com/1egoman))
+
+- update observability codegen - [#1335](https://github.com/livekit/protocol/pull/1335) ([@paulwe](https://github.com/paulwe))
+
+- Option to control auto subscribe of data tracks. - [#1365](https://github.com/livekit/protocol/pull/1365) ([@boks1971](https://github.com/boks1971))
+
+- Add project ID to internal SIPCall info. - [#1346](https://github.com/livekit/protocol/pull/1346) ([@dennwc](https://github.com/dennwc))
+
+- feat(whatsapp): add disconnect reason field for disconnect request - [#1341](https://github.com/livekit/protocol/pull/1341) ([@anunaym14](https://github.com/anunaym14))
+
+- redact egress assume_role_external_id in logs - [#1337](https://github.com/livekit/protocol/pull/1337) ([@paulwe](https://github.com/paulwe))
+
+- temporarily move extension ids to avoid etcd conflict - [#1352](https://github.com/livekit/protocol/pull/1352) ([@paulwe](https://github.com/paulwe))
+
+- Store repair ssrc in TrackInfo for migration purposes. - [#1348](https://github.com/livekit/protocol/pull/1348) ([@boks1971](https://github.com/boks1971))
+
+## 1.43.4
+
+### Patch Changes
+
+- Use string constant as parameter to twirp.NewErrorf - [#1326](https://github.com/livekit/protocol/pull/1326) ([@biglittlebigben](https://github.com/biglittlebigben))
+
+- add tags to agent recording proto - [#1328](https://github.com/livekit/protocol/pull/1328) ([@paulwe](https://github.com/paulwe))
+
+## 1.43.3
+
+### Patch Changes
+
+- Data tracks initial protocol definition. - [#1327](https://github.com/livekit/protocol/pull/1327) ([@boks1971](https://github.com/boks1971))
+
+- Cancellation counts for join/publish/subscribe. - [#1321](https://github.com/livekit/protocol/pull/1321) ([@boks1971](https://github.com/boks1971))
+
+- Add ai coustics feature usage - [#1324](https://github.com/livekit/protocol/pull/1324) ([@lukasIO](https://github.com/lukasIO))
+
+## 1.43.2
+
+### Patch Changes
+
+- fix call depth for override logger - [#1313](https://github.com/livekit/protocol/pull/1313) ([@paulwe](https://github.com/paulwe))
+
+- Generate agent protos - [#1318](https://github.com/livekit/protocol/pull/1318) ([@toubatbrian](https://github.com/toubatbrian))
+
+- add missing proto import - [#1317](https://github.com/livekit/protocol/pull/1317) ([@paulwe](https://github.com/paulwe))
+
+## 1.43.1
+
+### Patch Changes
+
+- add session features to room observability - [#1298](https://github.com/livekit/protocol/pull/1298) ([@paulwe](https://github.com/paulwe))
+
+- Redact metadata, attribute, participant name from logs. - [#1309](https://github.com/livekit/protocol/pull/1309) ([@boks1971](https://github.com/boks1971))
+
+- update room observability codegen - [#1307](https://github.com/livekit/protocol/pull/1307) ([@paulwe](https://github.com/paulwe))
+
+- Add redacted field options to egress fields - [#1286](https://github.com/livekit/protocol/pull/1286) ([@biglittlebigben](https://github.com/biglittlebigben))
+
+- update psrpc - [#1303](https://github.com/livekit/protocol/pull/1303) ([@paulwe](https://github.com/paulwe))
+
+- feat(connector): export protos from JS package - [#1311](https://github.com/livekit/protocol/pull/1311) ([@anunaym14](https://github.com/anunaym14))
+
+- add test for client middleware options - [#1306](https://github.com/livekit/protocol/pull/1306) ([@paulwe](https://github.com/paulwe))
+
+- add redact format option - [#1308](https://github.com/livekit/protocol/pull/1308) ([@paulwe](https://github.com/paulwe))
+
 ## 1.43.0
 
 ### Minor Changes